Refactor CommandPalette, CommandItem and CommandRowOnly components to move drag-and-drop functionality into CommandRowOnly
